requirements.txt
Flask
python-dotenv
scikit-learn
pandas
gunicorn
requests
